+++ example/spec/helpers/apiUtils.js
@@ -35,13 +35,22 @@ async function waitForApiRecord(getMethod, params, matchParams, retryConfig = {}
   );
 }
 
-async function throwIfApiError(apiFunction, apiFuncParams) {
-  const response = await apiFunction(apiFuncParams);
-  const parsedResponse = JSON.parse(response.body);
-  if (parsedResponse.error) {
-    throw new Error(parsedResponse.message);
-  }
-  return parsedResponse;
+async function waitForApiStatus(getMethod, params, status, retryConfig = {}) {
+  return await pRetry(
+    async () => {
+      const record = await getMethod(params);
+
+      const checkStatus = [status].flat();
+      if (!checkStatus.includes(record.status)) {
+        throw new Error(`Record status ${record.status}. Expect status ${status}`);
+      }
+      return record;
+    },
+    {
+      maxTimeout: 60 * 1000,
+      ...retryConfig,
+    }
+  );
 }
 
 /**
@@ -73,6 +82,6 @@ async function waitForModelStatus(model, params, status) {
 module.exports = {
   setDistributionApiEnvVars,
   waitForApiRecord,
+  waitForApiStatus,
   waitForModelStatus,
-  throwIfApiError,
 };