+++ example/spec/parallel/testAPI/createScheduledRuleSpec.js
@@ -7,12 +7,11 @@ const {
 } = require('@cumulus/integration-tests');
 
 const rulesApi = require('@cumulus/api-client/rules');
-const { deleteExecution } = require('@cumulus/api-client/executions');
+const { deleteExecution, getExecution } = require('@cumulus/api-client/executions');
 
 const {
-  throwIfApiError,
+  waitForApiStatus,
 } = require('../../helpers/apiUtils');
-
 const {
   createTestSuffix,
   createTimestampedTestId,
@@ -58,7 +57,7 @@ describe('When I create a scheduled rule via the Cumulus API', () => {
         testSuffix, testId);
       // Create a scheduled rule
       console.log(`creating rule ${scheduledRuleName}`);
-      await throwIfApiError(rulesApi.postRule, {
+      await rulesApi.postRule({
         prefix: config.stackName,
         rule: scheduledHelloWorldRule,
       });
@@ -68,6 +67,11 @@ describe('When I create a scheduled rule via the Cumulus API', () => {
   });
 
   afterAll(async () => {
+    await waitForApiStatus(
+      getExecution,
+      { prefix: config.stackName, arn: executionArn },
+      'completed'
+    );
     await deleteExecution({ prefix: config.stackName, executionArn });
     await cleanupCollections(config.stackName, config.bucket, collectionsDir,
       testSuffix);
@@ -107,7 +111,7 @@ describe('When I create a scheduled rule via the Cumulus API', () => {
       try {
         console.log(`deleting rule ${scheduledHelloWorldRule.name}`);
 
-        await throwIfApiError(rulesApi.deleteRule, {
+        await rulesApi.deleteRule({
           prefix: config.stackName,
           ruleName: scheduledHelloWorldRule.name,
         });