+++ packages/api/tests/endpoints/test-distribution.js
@@ -9,7 +9,7 @@ const { Cookie } = require('tough-cookie');
 const { URL } = require('url');
 const moment = require('moment');
 
-const { createBucket, s3PutObject, recursivelyDeleteS3Bucket } = require('@cumulus/aws-client/S3');
+const { createBucket, s3ObjectExists, s3PutObject, recursivelyDeleteS3Bucket } = require('@cumulus/aws-client/S3');
 const { RecordDoesNotExist } = require('@cumulus/errors');
 const { s3 } = require('@cumulus/aws-client/services');
 const { getLocalstackEndpoint } = require('@cumulus/aws-client/test-utils');
@@ -33,10 +33,14 @@ let headObjectStub;
 
 // import the express app after setting the env variables
 const { distributionApp } = require('../../app/distribution');
+const { writeBucketMapCacheToS3 } = require('../../endpoints/distribution');
 
 const publicBucket = randomId('publicbucket');
 const publicBucketPath = randomId('publicpath');
 const protectedBucket = randomId('protectedbucket');
+
+const headerContentTypeTextPlain = 'text/plain';
+const headerCustomHeaderVal = 'example-custom-header-value';
 
 const bucketMap = {
   MAP: {
@@ -46,9 +50,16 @@ const bucketMap = {
         'Content-Type': 'text/plain',
       },
     },
+    path2: {
+      bucket: 'bucket-path-2',
+      headers: {
+        'Content-Type': headerContentTypeTextPlain,
+        'Custom-Header': headerCustomHeaderVal,
+      },
+    },
+    anotherPath1: 'bucket-path-1',
     [protectedBucket]: protectedBucket,
     [publicBucketPath]: publicBucket,
-    [publicBucket]: publicBucket,
   },
   PUBLIC_BUCKETS: {
     [publicBucket]: 'public bucket',
@@ -309,6 +320,31 @@ test.serial('A request for a public file with an access token returns a redirect
   t.is(redirectLocation.searchParams.get('A-userid'), accessTokenRecord.username);
 });
 
+test.serial('A request for a public file with an access token in the bucket-map with headers defined in the bucket map returns a redirect to S3 containing the expected parameters to override the response headers', async (t) => {
+  stubHeadObject();
+
+  const { accessTokenCookie, accessTokenRecord, fileKey, s3Endpoint } = context;
+  const fileLocation = `bucket-path-2/${fileKey}`;
+  const response = await request(distributionApp)
+    .get(`/path2/${fileKey}`)
+    .set('Accept', 'application/json')
+    .set('Cookie', [`accessToken=${accessTokenCookie}`])
+    .expect(307);
+
+  restoreHeadObjectStub();
+
+  validateDefaultHeaders(t, response);
+
+  const redirectLocation = new URL(response.headers.location);
+
+  const signedFileUrl = new URL(`${s3Endpoint}/${fileLocation}`);
+  t.is(redirectLocation.origin, signedFileUrl.origin);
+  t.is(redirectLocation.pathname, signedFileUrl.pathname);
+  t.is(redirectLocation.searchParams.get('A-userid'), accessTokenRecord.username);
+  t.true(response.headers['content-type'].includes(headerContentTypeTextPlain));
+  t.is(response.headers['custom-header'], headerCustomHeaderVal);
+});
+
 test.serial('A /login request with a good authorization code returns a correct response', async (t) => {
   const {
     authorizationCode,
@@ -412,7 +448,7 @@ test.serial('A HEAD request for a public file without an access token redirects 
   const { fileKey, s3Endpoint } = context;
   const fileLocation = `${publicBucket}/${fileKey}`;
   const response = await request(distributionApp)
-    .head(`/${fileLocation}`)
+    .head(`/${publicBucketPath}/${fileKey}`)
     .set('Accept', 'application/json')
     .expect(307);
 
@@ -475,7 +511,7 @@ test('A sucessful /locate request for a bucket returns matching paths', async (t
     .set('Accept', 'application/json')
     .expect('Content-Type', /application\/json/)
     .expect(200);
-  t.deepEqual(response.body, ['path1']);
+  t.deepEqual(response.body, ['path1', 'anotherPath1']);
 });
 
 test('A /locate request returns error when no matching bucket found', async (t) => {
@@ -494,4 +530,41 @@ test('A /locate request returns error when request parameter is missing', async 
     .expect('Content-Type', /text\/plain/)
     .expect(400);
   t.true(JSON.stringify(response.error).includes('Required \\"bucket_name\\" query paramater not specified'));
+});
+
+test('writeBucketMapCacheToS3 creates distribution map cache in s3', async (t) => {
+  const bucketList = [protectedBucket, publicBucket];
+  const s3Bucket = process.env.system_bucket;
+  const s3Key = randomId('distributionBucketMap');
+
+  const mapObject = await writeBucketMapCacheToS3({ bucketList, s3Bucket, s3Key });
+  t.is(mapObject[protectedBucket], protectedBucket);
+  t.is(mapObject[publicBucket], publicBucketPath);
+  t.true(await s3ObjectExists({ Bucket: s3Bucket, Key: s3Key }));
+});
+
+test('writeBucketMapCacheToS3 throws error when there is no bucket mapping found for bucket', async (t) => {
+  const bucketName = randomId('nonexistPath');
+  const bucketList = [bucketName];
+  const s3Bucket = process.env.system_bucket;
+  const s3Key = randomId('distributionBucketMap');
+
+  try {
+    await writeBucketMapCacheToS3({ bucketList, s3Bucket, s3Key });
+  } catch (error) {
+    t.is(error.message, `No bucket mapping found for ${bucketName}`);
+  }
+});
+
+test('writeBucketMapCacheToS3 throws error when there are multiple paths found for bucket', async (t) => {
+  const bucketName = 'bucket-path-1';
+  const bucketList = [bucketName];
+  const s3Bucket = process.env.system_bucket;
+  const s3Key = randomId('distributionBucketMap');
+
+  try {
+    await writeBucketMapCacheToS3({ bucketList, s3Bucket, s3Key });
+  } catch (error) {
+    t.true(error.message.startsWith(`BucketMap configured with multiple responses from ${bucketName}`));
+  }
 });