+++ CHANGELOG.md
@@ -6,13 +6,41 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/).
 
 ## [Unreleased]
 
+### BREAKING CHANGES
+
+- All API requests made by `@cumulus/api-client` will now throw an error if the status code
+does not match the expected response (200 for most requests and 202 for a few requests that
+trigger async operations). Previously the helpers in this package would return the response
+regardless of the status code, so you may need to update any code using helpers from this
+package to catch or to otherwise handle errors that you may encounter.
+- The Cumulus API Lambda function has now been configured with reserved concurrency to ensure
+availability in a high-concurrency environment. However, this also caps max concurrency which
+may result in throttling errors if trying to reach the Cumulus API multiple times in a short
+period. Reserved concurrency can be configured with the `archive_api_reserved_concurrency`
+terraform variable on the Cumulus module and increased if you are seeing throttling erorrs.
+The default reserved concurrency value is 8.
+
 ### Added
 
 - Added user doc describing new features related to the Cumulus dead letter archive.
+- **CUMULUS-2327**
+  - Added reserved concurrency setting to the Cumulus API lambda function.
+  - Added relevant tfvars to the archive and cumulus terraform modules.
+- **CUMULUS-2460**
+  - Adds `POST` /executions/search-by-granules for retrieving executions from a list of granules or granule query
+  - Adds `searchExecutionsByGranules` to `@cumulus/api-client/executions`
 - **CUMULUS-2475**
   - Adds `GET` endpoint to distribution API
+- **CUMULUS-2463**
+  - `PUT /granules` reingest action allows a user to override the default execution
+    to use by providing an optional `workflowName` or `executionArn` parameter on
+    the request body.
+  - `PUT /granules/bulkReingest` action allows a user to override the default
+    execution/workflow combination to reingest with by providing an optional
+    `workflowName` on the request body.
+- Adds `workflowName` and `executionArn` params to @cumulus/api-client/reingestGranules
 - **CUMULUS-2476**
-  - Adds handler for authenticated `HEAD` Distribution requests replicating current behavior of TEA 
+  - Adds handler for authenticated `HEAD` Distribution requests replicating current behavior of TEA
 - **CUMULUS-2478**
   - Implemented [bucket map](https://github.com/asfadmin/thin-egress-app#bucket-mapping).
   - Implemented /locate endpoint
@@ -22,8 +50,90 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/).
     - (EDL only) restricts download from PRIVATE_BUCKETS to users who belong to certain EDL User Groups
     - bucket prefix and object prefix are supported
   - Add 'Bearer token' support as an authorization method
+- **CUMULUS-2486**
+  - Implemented support for custom headers
+  - Added 'Bearer token' support as an authorization method
 - **CUMULUS-2487**
-  - Add integration test for cumulus distribution API
+  - Added integration test for cumulus distribution API
+- **CUMULUS-2569**
+  - Created bucket map cache for cumulus distribution API
+- **CUMULUS-2568**
+  - Add `deletePdr`/PDR deletion functionality to `@cumulus/api-client/pdrs`
+  - Add `removeCollectionAndAllDependencies` to integration test helpers
+  - Added `example/spec/apiUtils.waitForApiStatus` to wait for a
+  record to be returned by the API with a specific value for
+  `status`
+  - Added `example/spec/discoverUtils.uploadS3GranuleDataForDiscovery` to upload granule data fixtures
+  to S3 with a randomized granule ID for `discover-granules` based
+  integration tests
+  - Added `example/spec/Collections.removeCollectionAndAllDependencies` to remove a collection and
+  all dependent objects (e.g. PDRs, granules, executions) from the
+  database via the API
+  - Added helpers to `@cumulus/api-client`:
+    - `pdrs.deletePdr` - Delete a PDR via the API
+    - `replays.postKinesisReplays` - Submit a POST request to the `/replays` endpoint for replaying Kinesis messages
+
+### Changed
+
+- Moved functions from `@cumulus/integration-tests` to `example/spec/helpers/workflowUtils`:
+  - `startWorkflowExecution`
+  - `startWorkflow`
+  - `executeWorkflow`
+  - `buildWorkflow`
+  - `testWorkflow`
+  - `buildAndExecuteWorkflow`
+  - `buildAndStartWorkflow`
+- `example/spec/helpers/workflowUtils.executeWorkflow` now uses
+`waitForApiStatus` to ensure that the execution is `completed` or
+`failed` before resolving
+- `example/spec/helpers/testUtils.updateAndUploadTestFileToBucket`
+now accepts an object of parameters rather than positional
+arguments
+- Removed PDR from the `payload` in the input payload test fixture for reconciliation report integration tests
+- The following integration tests for PDR-based workflows were
+updated to use randomized granule IDs:
+  - `example/spec/parallel/ingest/ingestFromPdrSpec.js`
+  - `example/spec/parallel/ingest/ingestFromPdrWithChildWorkflowMetaSpec.js`
+  - `example/spec/parallel/ingest/ingestFromPdrWithExecutionNamePrefixSpec.js`
+  - `example/spec/parallel/ingest/ingestPdrWithNodeNameSpec.js`
+- Updated the `@cumulus/api-client/CumulusApiClientError` error class to include new properties that can be accessed directly on
+the error object:
+  - `statusCode` - The HTTP status code of the API response
+  - `apiMessage` - The message from the API response
+- Added `params.pRetryOptions` parameter to
+`@cumulus/api-client/granules.deleteGranule` to control the retry
+behavior
+- **CUMULUS-2482**
+  - Switches the default distribution app in the `example/cumulus-tf` deployment to the new Cumulus Distribution
+  - TEA is still available by following instructions in `example/README.md`
+- **CUMULUS-2463**
+  - Increases the duration of allowed backoff times for a successful test from 0.5 sec to 1 sec.
+
+
+### Fixed
+
+- Fixed bug where `cmr_custom_host` variable was not properly forwarded into `archive`, `ingest`, and `sqs-message-remover` modules from `cumulus` module
+- Fixed bug where `parse-pdr` set a granule's provider to the entire provider record when a `NODE_NAME`
+  is present. Expected behavior consistent with other tasks is to set the provider name in that field.
+- **CUMULUS-2568**
+  - Update reconciliation report integration test to have better cleanup/failure behavior
+  - Fixed `@cumulus/api-client/pdrs.getPdr` to request correct endpoint for returning a PDR from the API
+- **CUMULUS-2620**
+  - Fixed a bug where a granule could be removed from CMR but still be set as
+  `published: true` and with a CMR link in the Dynamo/PostgreSQL databases. Now,
+  the CMR deletion and the Dynamo/PostgreSQL record updates will all succeed or fail
+  together, preventing the database records from being out of sync with CMR.
+  - Fixed `@cumulus/api-client/pdrs.getPdr` to request correct
+  endpoint for returning a PDR from the API
+
+## [v9.2.0] 2021-06-22
+
+### Added
+
+- **CUMULUS-2475**
+  - Adds `GET` endpoint to distribution API
+- **CUMULUS-2476**
+  - Adds handler for authenticated `HEAD` Distribution requests replicating current behavior of TEA
 
 ### Changed
 - **CUMULUS-2482**
@@ -4448,7 +4558,8 @@ Note: There was an issue publishing 1.12.0. Upgrade to 1.12.1.
 
 ## [v1.0.0] - 2018-02-23
 
-[unreleased]: https://github.com/nasa/cumulus/compare/v9.1.0...HEAD
+[unreleased]: https://github.com/nasa/cumulus/compare/v9.2.0...HEAD
+[v9.2.0]: https://github.com/nasa/cumulus/compare/v9.1.0...v9.2.0
 [v9.1.0]: https://github.com/nasa/cumulus/compare/v9.0.1...v9.1.0
 [v9.0.1]: https://github.com/nasa/cumulus/compare/v9.0.0...v9.0.1
 [v9.0.0]: https://github.com/nasa/cumulus/compare/v8.1.0...v9.0.0